requirements.txt
python-telegram-bot==20.3
requests
psutil
beautifulsoup4
lxml